+++ apps/src/codebridge/hooks/useZoomTracker.ts
@@ -1,7 +1,7 @@
-import {sendLab2AnalyticsEvent} from '@codebridge/utils/analyticsReporterHelper';
 import debounce from 'lodash/debounce';
 import {useEffect, useRef} from 'react';
 
+import {sendLab2AnalyticsEvent} from '@cdo/apps/lab2/utils/analyticsReporterHelper';
 import {EVENTS} from '@cdo/apps/metrics/AnalyticsConstants';
 
 const DEBOUNCE_TIMEOUT = 1500;
