+++ apps/src/lab2/utils/analyticsReporterHelper.ts
